exchange complete

+++ resources/views/sales/home.blade.php
@@ -53,7 +53,7 @@
                         <td>
                             <a class="btn btn-success" href="{{route('sales.edit', $data->id)}}">Due</a>
                             <a class="btn btn-warning" href="{{route('sales.show', $data->id)}}">Exch.</a>
-                            {!!Form::open(array('route' => ['sales.destroy', $data->id], 'class'=>'pull-right', 'method'=>'delete'))!!}
+                            {!!Form::open(array('route' => ['sales.destroy', $data->id], 'class'=>'pull-right', 'method'=>'delete', 'onclick'=>'return confirm("Are you sure you want to delete this item?");'))!!}
                             {!!Form::hidden('invoice_no', $data->invoice_no)!!}
                             {!!Form::hidden('products', $data->products)!!}
                             {!!Form::hidden('quantity', $data->quantity)!!}
